Move fixtures to confeste.py
